# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flappy Wif Hat game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUrl {
  url: string;
}

export interface User {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  bthCoins: number;
  highScore: number;
  isGuest: boolean;
  /** @nullable */
  twitterId?: string | null;
}

export interface GuestLoginInput {
  /**
     * @minLength 1
     * @maxLength 20
     */
  username: string;
}

export interface ScoreInput {
  /** @minimum 0 */
  score: number;
}

export interface ScoreResult {
  score: number;
  bthCoinsEarned: number;
  totalBthCoins: number;
  isHighScore: boolean;
  /** @nullable */
  rank?: number | null;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  highScore: number;
  bthCoins: number;
  isGuest: boolean;
}

export interface UserProfile {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  bthCoins: number;
  highScore: number;
  isGuest: boolean;
  gamesPlayed: number;
  /** @nullable */
  rank?: number | null;
}

export interface ErrorResponse {
  error: string;
}

export type TwitterCallbackParams = {
code?: string;
state?: string;
oauth_token?: string;
oauth_verifier?: string;
};

export type GetLeaderboardParams = {
limit?: number;
};