# lib/api-zod/src/generated/types/guestLoginInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flappy Wif Hat game API
 * OpenAPI spec version: 0.1.0
 */

export interface GuestLoginInput {
  /**
     * @minLength 1
     * @maxLength 20
     */
  username: string;
}