# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flappy Wif Hat game API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  highScore: number;
  bthCoins: number;
  isGuest: boolean;
}