# lib/api-zod/src/generated/types/scoreInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flappy Wif Hat game API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreInput {
  /** @minimum 0 */
  score: number;
}