# lib/api-zod/src/generated/types/scoreResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flappy Wif Hat game API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreResult {
  score: number;
  bthCoinsEarned: number;
  totalBthCoins: number;
  isHighScore: boolean;
  /** @nullable */
  rank?: number | null;
}