# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flappy Wif Hat game API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  bthCoins: number;
  highScore: number;
  isGuest: boolean;
  /** @nullable */
  twitterId?: string | null;
}